# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Invoice Generator API",
...
    "version": "1.0.0",
    "summary": "Professional Invoice Generator API",
    "contact": {
      "name": "Invoice Generator API Support Team",
      "email": "support@yourdomain.com",
      "url": "https://yourdomain.com/contact"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://yourdomain.com/terms"
  },
  "servers": [
    {
      "url": "https://screeching-tildi-adelzidoune-ca9a7151.koyeb.app",
      "description": "Koyeb Production Server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "description": "Create a new user account\n\n**No authentication required**\n\nCreates a new user with:\n- Unique email and username\n- Hashed password (bcrypt)\n- Optional company information\n- Automatic payment link generation",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "username", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "freelancer@example.com",
                    "description": "Valid email address"
                  },
                  "username": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 50,
                    "example": "freelancer",
                    "description": "Unique username (3-50 characters)"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "minLength": 8,
                    "example": "securepass123",
                    "description": "Strong password (min 8 characters)"
                  },
                  "full_name": {
                    "type": "string",
                    "example": "Ahmed Freelancer",
                    "description": "Full name (optional)"
                  },
                  "company_name": {
                    "type": "string",
                    "example": "Freelance Pro",
                    "description": "Company name (optional)"
                  },
                  "phone": {
                    "type": "string",
                    "example": "+212600000000",
                    "description": "Phone number (optional)"
                  },
                  "address": {
                    "type": "string",
                    "example": "123 Street, Casablanca, Morocco",
                    "description": "Full address (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string", "example": "User created successfully"},
                    "user_id": {"type": "integer", "example": 1},
                    "username": {"type": "string", "example": "freelancer"}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Email or username already exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {"type": "string", "example": "Email already registered"}
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "loc": {"type": "array"},
                          "msg": {"type": "string"},
                          "type": {"type": "string"}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login user",
        "description": "Authenticate user and return JWT token\n\n**No authentication required**\n\nReturns:\n- JWT access token (valid for 30 minutes)\n- Token type (Bearer)\n- User information",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "freelancer",
                    "description": "Username or email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "securepass123",
                    "description": "User password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    },
                    "token_type": {
                      "type": "string",
                      "example": "bearer"
                    },
                    "expires_in": {
                      "type": "integer",
                      "example": 1800,
                      "description": "Token expiration time in seconds"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {"type": "string", "example": "Invalid username or password"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/invoices/generate": {
      "post": {
        "tags": ["Invoices"],
        "summary": "Create a new invoice",
        "description": "Generate a new invoice with items and calculations\n\n**Authentication required**\n\nCreates:\n- Bilingual PDF (Arabic/English)\n- QR code with payment link\n- Invoice number (auto-generated)\n- Tax and discount calculations\n\n**Supported currencies**: MAD, USD, EUR, SAR, AED, GBP, EGP",
        "operationId": "generate_invoice",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["client_name", "client_email", "items"],
                "properties": {
                  "client_name": {
                    "type": "string",
                    "example": "ACME Corporation",
                    "description": "Client company name"
                  },
                  "client_email": {
                    "type": "string",
                    "format": "email",
                    "example": "billing@acme.com",
                    "description": "Client email for invoice delivery"
                  },
                  "client_phone": {
                    "type": "string",
                    "example": "+1234567890",
                    "description": "Client phone number (optional)"
                  },
                  "client_address": {
                    "type": "string",
                    "example": "123 Business Ave, New York, USA",
                    "description": "Client full address (optional)"
                  },
                  "language": {
                    "type": "string",
                    "enum": ["en", "ar", "fr"],
                    "default": "en",
                    "example": "ar",
                    "description": "Invoice language (en=English, ar=Arabic, fr=French)"
                  },
                  "currency": {
                    "type": "string",
                    "enum": ["MAD", "USD", "EUR", "SAR", "AED", "GBP", "EGP"],
                    "default": "USD",
                    "example": "MAD",
                    "description": "Currency code"
                  },
                  "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                      "type": "object",
                      "required": ["name", "quantity", "price"],
                      "properties": {
                        "name": {
                          "type": "string",
                          "example": "تطوير موقع ويب",
                          "description": "Item/service name"
                        },
                        "description": {
                          "type": "string",
                          "example": "موقع تجارة إلكترونية كامل",
                          "description": "Detailed description (optional)"
                        },
                        "quantity": {
                          "type": "number",
                          "minimum": 0.01,
                          "example": 1,
                          "description": "Quantity or hours"
                        },
                        "price": {
                          "type": "number",
                          "minimum": 0,
                          "example": 15000,
                          "description": "Unit price"
                        }
                      }
                    },
                    "description": "List of invoice items (min 1 item)"
                  },
                  "tax_rate": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100,
                    "default": 0,
                    "example": 20,
                    "description": "Tax rate percentage (0-100)"
                  },
                  "discount_rate": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100,
                    "default": 0,
                    "example": 10,
                    "description": "Discount rate percentage (0-100)"
                  },
                  "due_date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-11-10T00:00:00",
                    "description": "Payment due date (optional)"
                  },
                  "notes": {
                    "type": "string",
                    "example": "شكراً لتعاملكم معنا. الدفع خلال 30 يوم",
                    "description": "Additional notes or payment terms (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invoice created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {"type": "integer", "example": 1},
                    "invoice_number": {"type": "string", "example": "INV-2025-0001"},
                    "status": {"type": "string", "example": "draft"},
                    "total": {"type": "number", "example": 16200},
                    "pdf_url": {"type": "string", "example": "/static/invoices/INV-2025-0001.pdf"},
                    "qr_code_url": {"type": "string", "example": "/static/qr_codes/INV-2025-0001.png"}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized - Invalid or expired token"
          },
          "422": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/invoices/{invoice_id}/download": {
      "get": {
        "tags": ["Invoices"],
        "summary": "Download invoice PDF",
        "description": "Download invoice as PDF file\n\n**Authentication required**\n\nReturns the invoice PDF file for download.",
        "operationId": "download_invoice",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF file",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Invoice not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme.\n\n**How to use:**\n1. Login at `/auth/login`\n2. Copy the `access_token` from response\n3. Click 🔓 **Authorize** button above\n4. Enter: `Bearer <your_token>`\n5. Click **Authorize**\n\n**Token lifetime:** 30 minutes\n\n**Example:**\n```\nAuthorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\n```"
      }
    },
...
  },
  "security": [],
...
}